requirements.txt
torch>=1.11
torchvision>=0.12
numpy>=1.19
tqdm>=4.62
tensorboard>=2.0
matplotlib>=3.3
Pillow>=8.0
opencv-python>=4.5
scikit-learn>=1.0
